Started Networking Features; a lot of deleting and creating new game objects and components
The wheels are not able to rotate because they are not assigned to an
instance of an object; there are no trucks until the game begins, and I
cannot assign movement before. One of the trucks is not facing the
correct direction on spawn; tried to set the spawn location scale to a
negative x value…. The trucks do not re-spawn when the game resets. Was
using two new truck scripts to reset the position and calling this
method from the GameManager on reset; this no longer is viable.